Move the mocked code to mock.rs
