+++ runtime/calamari/tests/common/mod.rs
@@ -14,21 +14,12 @@
 // You should have received a copy of the GNU General Public License
 // along with Manta.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Calamari Parachain Integration Tests.
+pub mod mock;
 
-pub use calamari_runtime::{
-	currency::KMA, Call, CollatorSelection, Democracy, Event, Origin, Runtime, Scheduler, Session,
-	System, TransactionPayment,
-};
-use frame_support::{
-	traits::{GenesisBuild, OnFinalize, OnInitialize},
-	weights::{DispatchInfo, Weight},
-};
-use manta_primitives::{
-	helpers::{get_account_id_from_seed, get_collator_keys_from_seed},
-	AccountId, AuraId, Balance,
-};
-use sp_core::sr25519;
+pub use calamari_runtime::{currency::KMA, Event, Origin, Runtime, System};
+
+use frame_support::weights::{DispatchInfo, Weight};
+use manta_primitives::Balance;
 
 pub const BOND_AMOUNT: Balance = 1_000 * KMA;
 pub const INITIAL_BALANCE: Balance = 1_000_000_000_000 * KMA;
@@ -42,121 +33,8 @@ pub fn info_from_weight(w: Weight) -> DispatchInfo {
 	}
 }
 
-/// Utility function that advances the chain to the desired block number.
-pub fn run_to_block(n: u32) {
-	while System::block_number() < n {
-		System::on_finalize(System::block_number());
-		System::set_block_number(System::block_number() + 1);
-		System::on_initialize(System::block_number());
-		Session::on_initialize(System::block_number());
-		Scheduler::on_initialize(System::block_number());
-		Democracy::on_initialize(System::block_number());
-	}
-}
-
 pub fn last_event() -> Event {
 	System::events().pop().expect("Event expected").event
-}
-
-pub struct ExtBuilder {
-	balances: Vec<(AccountId, Balance)>,
-	authorities: Vec<(AccountId, AuraId)>,
-	invulnerables: Vec<AccountId>,
-	desired_candidates: u32,
-	safe_xcm_version: Option<u32>,
-}
-
-impl Default for ExtBuilder {
-	fn default() -> ExtBuilder {
-		ExtBuilder {
-			balances: vec![(
-				get_account_id_from_seed::<sr25519::Public>("Alice"),
-				INITIAL_BALANCE,
-			)],
-			authorities: vec![(
-				get_account_id_from_seed::<sr25519::Public>("Alice"),
-				get_collator_keys_from_seed("Alice"),
-			)],
-			invulnerables: vec![get_account_id_from_seed::<sr25519::Public>("Alice")],
-			safe_xcm_version: None,
-			desired_candidates: 1,
-		}
-	}
-}
-
-impl ExtBuilder {
-	pub fn with_balances(mut self, balances: Vec<(AccountId, Balance)>) -> Self {
-		self.balances = balances;
-		self
-	}
-
-	pub fn with_authorities(mut self, authorities: Vec<(AccountId, AuraId)>) -> Self {
-		self.authorities = authorities;
-		self
-	}
-
-	pub fn with_collators(
-		mut self,
-		invulnerables: Vec<AccountId>,
-		desired_candidates: u32,
-	) -> Self {
-		self.invulnerables = invulnerables;
-		self.desired_candidates = desired_candidates;
-		self
-	}
-
-	pub fn build(self) -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::default()
-			.build_storage::<Runtime>()
-			.unwrap();
-
-		pallet_balances::GenesisConfig::<Runtime> {
-			balances: self.balances,
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		manta_collator_selection::GenesisConfig::<Runtime> {
-			invulnerables: self.invulnerables,
-			candidacy_bond: BOND_AMOUNT,
-			desired_candidates: self.desired_candidates,
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		pallet_session::GenesisConfig::<Runtime> {
-			keys: self
-				.authorities
-				.iter()
-				.cloned()
-				.map(|(acc, aura)| {
-					(
-						acc.clone(),                                    // account id
-						acc,                                            // validator id
-						calamari_runtime::opaque::SessionKeys { aura }, // session keys
-					)
-				})
-				.collect(),
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		<pallet_xcm::GenesisConfig as GenesisBuild<Runtime>>::assimilate_storage(
-			&pallet_xcm::GenesisConfig {
-				safe_xcm_version: self.safe_xcm_version,
-			},
-			&mut t,
-		)
-		.unwrap();
-
-		let mut ext = sp_io::TestExternalities::new(t);
-
-		ext.execute_with(|| {
-			System::set_block_number(1);
-		});
-
-		ext
-	}
 }
 
 pub fn root_origin() -> <Runtime as frame_system::Config>::Origin {
